Moved assets to the theme's root.
There are a couple of ways to go when architecting assets:

1. Put the asset files in the theme's root.
2. Create an assets module where everything associated with the assets
is handled from that module.

-----------------------
Option 1: Root
-----------------------

With this option, assets are served up from the root of the theme. The
folder then holds the theme's assets, but not the business logic to
enqueue or serve them up.

When to use this strategy?

Use this strategy when the assets are general to the theme and
NOT part of a module or submodule.

Advantages to this strategy are:

1. each asset's URL has one less folder in it.
2. assets are not contained in the business logic.
3. allows for images, fonts, svgs, or other assets to be packaged with the theme.
4. this folder houses the custom elements of your child theme, those assets that
will change from project-to-project.
5. the business logic will be more portable and reusable from project-to-project.

-----------------------
Option 2: Assets Module
-----------------------

For this option, the assets and everything associated with them is
grouped together into a module.  That means, you'd move the enqueuing
tasks into this module.

When to use this strategy?

Use this strategy when you to group everything associated with assets
into a single module or when it's part of a submodule.

Advantages for this strategy are:

1. The business logic and the assets are in a module.

Disadvantages are:

1. The module contains custom code that is not reusable from project-to-project.
2. When adding images, svgs, etc. into the module, structural views and logic
will need to be moved to this module.  This may lead to some confusion.