Fixed the favicon.
This commit adds the missing favicon images to the assets folder.

It also changes the rendering of the favicon logic.  Beans has favicon
rendering logic.  We can hook into the attributes and change the `href` of
the link.  Using this strategy, we can leverage Beans and simply redirect
the link to where we store images, i.e. in `assets/images`.