Switched to string-based text domain.
Many internationalization apps will not process code.  Therefore, we need to use
a native string for the the text domain, as the app won't convert the constant.
Of course that means every single theme has now be customized to change the text
domain. Argh! Frustrating, I know!  Psst...working on a solution for this.